% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc-methods.R
\name{calcBalance}
\alias{calcBalance}
\title{Calculate balance of tree}
\usage{
calcBalance(tree, node = "all")
}
\arguments{
\item{tree}{phylogenetic tree (ape class)}

\item{node}{which node to calculate balance from, default 'all'}
}
\description{
Return data.frame of nodal imbalance
}
\references{
No references yet!
\url{https://github.com/DomBennett/MoreTreeTools}.
}
\keyword{ecology}
\keyword{evolution}
\keyword{manipulation}
\keyword{palaeontology}
\keyword{phylogeny}

