% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clade-methods.R
\name{calcCladeStats}
\alias{calcCladeStats}
\title{Calculate clade stats}
\usage{
calcCladeStats(clades)
}
\arguments{
\item{clades}{dataframe where columns are clades and rows are numbers
of descendents by time step as returned by getCladeSuccess()}
}
\description{
Return a dataframe of stats for dataframe produced from getCladeSuccess().
}
\details{
The function calculates the following statistics for each clade:

tot.size: the total number of tips in the clade through time

max.size: the maximum number of tips in the clade in one time step

time.span: the number of time steps the clade existed

start: the starting time step

end: the ending time step

cm: the centre of mass of the clade

cg: the centre of gyration of the clade
}
\references{
Gould, S. J., Gilinsky, N. L., & German, R. Z. (1987). Asymmetry of
lineages and the direction of evolutionary time. Science (New York, N.Y.),
236(4807), 1437–1441.
}

