% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calc-methods.R
\name{calcComPhyMets}
\alias{calcComPhyMets}
\title{Calculate Community Phylogenetic Metrics}
\usage{
calcComPhyMets(cmatrix, tree, min.spp = 2, metrics = c("PD1", "PD2", "PD3",
  "PSV", "PSD", "PSE", "PSC"))
}
\arguments{
\item{cmatrix}{community/trait data matrix (cols taxa, rows sites)}

\item{tree}{phylogenetic tree (ape class)}

\item{min.spp}{minimum number of species at site, default 2}

\item{metric}{what metric to use, as a single character or vector, default all metrics}
}
\description{
One-stop function for calculating a range of community phylogenetic metrics
from tree and community matrix.
}
\details{
Metrics calculated: PD (types 1, 2 and 3), PSV, PSD, PSE, PSC, and PSR.
The PD metric in this function has 3 different types:
 PD1: considers the case if all other taxa from the phylogeny are dropped. In this
 way, type 1 is context independent but has a minimum number of taxa of 2. (This is
 the normal PD.)
 PD2: sums the lengths of all the edges connecting the specified taxa to the terminal node.
 PD3: sums the lengths of all the edges represented uniquely by the specified taxa.
}
\examples{
data (catarrhines)
 cmatrix <- randCommData (catarrhines, nsites=20, nspp=10)
 res <- calcComPhyMets (cmatrix, catarrhines)
 print (res)
}
\references{
No references yet!
\url{https://github.com/DomBennett/MoreTreeTools}.

Faith, D. (1992). Conservation evaluation and phylogenetic diversity.
 Biological Conservation, 61, 1–10.

Helmus M.R., Bland T.J., Williams C.K. & Ives A.R. (2007) Phylogenetic measures of
 biodiversity. American Naturalist, 169, E68-E83
}
\keyword{ecology}
\keyword{evolution}
\keyword{manipulation}
\keyword{palaeontology}
\keyword{phylogeny}

