% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calc-methods.R
\name{calcDist}
\alias{calcDist}
\title{Calculate the distance between trees using different methods}
\usage{
calcDist(tree1, tree2, method = c("all", "PH85", "score", "dmat"),
  normalised = TRUE)
}
\arguments{
\item{tree1}{first tree of comparison}

\item{tree2}{second tree of comparison}

\item{method}{'all', 'PH85', 'score' or 'dmat'. Default 'all'}

\item{normalised}{boolean, return distances of 0-1}
}
\description{
Calculate the normalised tree distance (topological and branch length)
between trees
}
\details{
This functions uses three different methods for calculating the distance
between trees: the 'PH85' and 'score' methods of ape's topo.dist and/or 1 - Pearson's r
calculated from the cophenetic matrix of the tip distances ('dmat'). The function returns
the normalised distances by default. Trees of different numbers of tips are shrunk
to the same size. Function will also rescale branch distances to sum to 1 before
calculating distances.
}

