% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calc-methods.R
\name{calcED}
\alias{calcED}
\title{Calculate Evolutionary Distinctness}
\usage{
calcED(tree, tips = "all", type = "FP")
}
\arguments{
\item{tree}{phylogenetic tree (ape class)}

\item{tips}{vector of tips for which ED is calculated, else 'all'}

\item{type}{method of ED calculation, either 'all', 'FP', 'ES' or 'PE' (default FP)}
}
\description{
Calculate species evolutionary distinctness (ED) using one of three methods:
 Fair Proportion (FP), Equal Splits (ES) or Pendant Edge (PE).
}
\details{
Evolutionary distinctness is a measure of how much independent evolution a species
 represents. Multiple methods exist for its calculation all of which require an ultrametric
 phylogenetic tree. The methods used here are Pendant Edge (PE) the length of a species branch
 that connects it to the tree (Altschul and Lipman, 1990), Fair Proportion (FP) the total
 proportion of the phylogenetic tree that a species represents where each branch is equally
 divided between all descdendants (Isaac et al. 2007) and Equal Splits (ES) where branch lengths
 are equally divided between descendents at every node in the tree (Redding and Mooers 2006)
}
\examples{
# Load catarrhine tree
data(catarrhines)
ed.vals <- calcED (catarrhines, type='all')
# Which Old World monkeys are the most distinct?
ed.vals[ed.vals$FP == min (ed.vals$FP), ]
ed.vals[ed.vals$ES == min (ed.vals$ES), ]
ed.vals[ed.vals$PE == min (ed.vals$PE), ]
}
\references{
No references yet!
\url{https://github.com/DomBennett/MoreTreeTools}.

Isaac, N.J.B., Turvey, S.T., Collen, B., Waterman, C. and Baillie, J.E.M. (2007).
 Mammals on the EDGE: conservation priorities based on threat and phylogeny. PLoS ONE, 2, e296.

 Altschul, S., and Lipman, D. (1990). Equal animals. Nature.

 Redding, D. W., and Mooers, A. Ø. (2006). Incorporating evolutionary measures into conservation
 prioritization. Conservation Biology : The Journal of the Society for Conservation Biology,
 20(6), 1670–8.
}
\seealso{
\code{\link[picante]{evol.distinct}}, \code{\link{catarrhines}}
}
\keyword{ecology}
\keyword{evolution}
\keyword{manipulation}
\keyword{palaeontology}
\keyword{phylogeny}

