% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calc-methods.R
\name{calcEdgeDiversity}
\alias{calcEdgeDiversity}
\title{Calculate the diversity of edges within set time intervals}
\usage{
calcEdgeDiversity(tree, n.intervals)
}
\arguments{
\item{tree}{phylogenetic tree (ape class)}

\item{n.intervals}{number of intervals}
}
\description{
Return a dataframe of edge number and a count
based on the number descendents within n.intervals uniformly
spaced time intervals.
}
\details{
Intervals are unifromely spaced from the root to the tip
of the tree. Output can be provided to treeplot for colouring edges,
see example.
}
\examples{
tree <- rtree (100)
edge.diversity <- calcEdgeDiversity (tree, n.intervals=4)
# convert to logged z-score to increase colour spectrum in plot
edge.diversity$count <- log (edge.diversity$count)
edge.diversity$col <- (edge.diversity$count - mean (edge.diversity$count)) / sd (edge.diversity$count)
treeplot (tree, edge.cols=edge.diversity, legend.title='Diversity')
}
\references{
No references yet!
\url{https://github.com/DomBennett/MoreTreeTools}.
}
\keyword{ecology}
\keyword{evolution}
\keyword{manipulation}
\keyword{palaeontology}
\keyword{phylogeny}

