\name{catarrhines}
\alias{catarrhines}
\docType{data}
\title{Phylogenetic tree of Catarrhini}
\description{
  Phylogenetic tree of the Parvorder Catarrhini, sister group of the New World Monkeys consisting of the Old World Monkeys and Apes, trimmed from Bininda-Edmonds et al. (2007)'s Mammalian supertree. Here used for testing and demonstrating MoreTreeTools functions.
}
\usage{
   data(catarrhines)
}
\format{
  \code{catarrhines} is a \code{phylo} from the pacakge \code{ape}.
}
\source{
   Bininda-Edmonds et al. (2007) The Delayed Rise of Present-Day Mammals. Nature, 446(7135): 507-512
}
\examples{
data (catarrhines) # load object
plot (catarrhines) # plot
}
\keyword{datasets}
