% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/build-methods.R
\name{collapseTips}
\alias{collapseTips}
\title{Remove all tips smaller than min.length}
\usage{
collapseTips(tree, min.length, iterative = TRUE)
}
\arguments{
\item{tree}{phylogenetic tree (ape class)}

\item{min.length}{smallest length for a tip edge}

\item{iterative}{run until no tip edges less than min.length, default TRUE}
}
\description{
Return a phylo object with tip edges less than min.length
removed.
}
\details{
Tips are removed, but the tree age is preserved. Returned
trees are therefore not the same phylogenetic tree as that given.
The purpose of the function is to generate representations of
trees with fewer tips for plotting purposes.


Set iterative as TRUE to keep removing all tip edges less than min.length.
}
\examples{
tree <- rtree (100)
par (mfrow = c (1,2))
plot (tree, show.tip.label=FALSE, main='Full tree')
tree.age <- getSize (tree, 'rtt')
ctree <- collapseTips (tree, min.length=tree.age*0.1)
plot (ctree, show.tip.label=FALSE, main='Collapsed tree')
}
\references{
No references yet!
\url{https://github.com/DomBennett/MoreTreeTools}.
}
\keyword{ecology}
\keyword{evolution}
\keyword{manipulation}
\keyword{palaeontology}
\keyword{phylogeny}

