% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viz-methods.R
\name{commplot}
\alias{commplot}
\title{Plot a community or traits on a phylogenetic tree}
\usage{
commplot(cmatrix, tree, groups = rep(1, nrow(cmatrix)), no.margin = TRUE,
  ...)
}
\arguments{
\item{cmatrix}{community/trait data matrix (cols taxa, rows sites)}

\item{tree}{phylogenetic tree (ape class)}

\item{groups}{trait or site groups}
}
\description{
Plot community or trait data on phylogeny to visualise
differences in structure between traits or sites. Use colours to distinguish
groups and alpha to distinguish abundances (works like rainbow())
}
\details{
No details
}
\examples{
# Generate a balanced tree
tree <- compute.brlen (stree (16, 'balanced'))
# Generate random community data, 2 of 3 different communities
cmatrix <- matrix (round (runif (16*6, 0, 2)), nrow = 6, byrow = TRUE)
colnames (cmatrix) <- tree$tip.label
rownames (cmatrix) <- paste0 ('site_', 1:6)
# Plot community, specify group identitiy of each of the 6 sites
commplot (cmatrix, tree, groups = c (1,1,2,2,3,3))
}
\references{
No references yet!
\url{https://github.com/DomBennett/MoreTreeTools}.
}
\keyword{ecology}
\keyword{evolution}
\keyword{manipulation}
\keyword{palaeontology}
\keyword{phylogeny}

