% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calc-methods.R
\name{extractTree}
\alias{extractTree}
\title{Extract tree for names given}
\usage{
extractTree(tree, names, fuzzy = TRUE, stats = FALSE)
}
\arguments{
\item{tree}{phylogenetic tree (ape class)}

\item{names}{vector of names of tips to be extracted}

\item{fuzzy}{boolean, if true will search Global Names Resolver online}

\item{stats}{boolean, if true returns list object of extracted tree and extraction statistics}
}
\description{
Return a tree of the names given with the option of searching online
to perform fuzzy-name matching.
}
\details{
This function firsts matches names to tip labels in tree, if not all names
are present in the tree and 'fuzzy' is true, it will then search the online taxonomic
names resolution service Global Names Resolver. NCBI taxonomic IDs are then matched
between tip labels and names, if names are still not mapped to tree, names are mapped
to tips using random mapping. For example, if an unmapped name is a member of the same
genus as several species in the tree, the name will be mapped to the tree with any one
of these at random.

If stats is true, the function will return a list of the extracted tree and a data frame
giving the proportions of names mapped to the tree, names mapped by exact string matching,
fuzzy matching and random taxonomic group placement. Of those names that are mapped using
random taxonomic placement, the mean NCBI rank number is returned. These are roughly
equivalent to:

Rank         Number

superkingdom    2

kingdom         4

phylum          9

subphylum       10

superclass      12

class           18

superorder      21

order           22

suborder        23

family          24

subfamily       25

genus           26

species         27
}
\references{
No references yet!
\url{https://github.com/DomBennett/MoreTreeTools}.
}
\keyword{ecology}
\keyword{evolution}
\keyword{manipulation}
\keyword{palaeontology}
\keyword{phylogeny}

