\name{getAge}
\alias{getAge}
\title{Return age of node given}
\usage{
getAge(tree, node)
}
\arguments{
  \item{tree}{phylogenetic tree (ape class)}

  \item{node}{number indicating node}
}
\description{
Return age of tip or internal node so long as the tree
provided is time calibrated with branch lengths.
}
\details{
No details
}
\examples{
# example.var <- exampleFun (test.data)
}
\references{
No references yet!
\url{https://github.com/DomBennett/MoreTreeTools}.
}
\keyword{ecology}
\keyword{evolution}
\keyword{manipulation}
\keyword{palaeontology}
\keyword{phylogeny}

