% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get-methods.R
\name{getClades}
\alias{getClades}
\title{Return all children of each node}
\usage{
getClades(tree)
}
\arguments{
\item{tree}{phylogenetic tree (ape class)}
}
\description{
All children for each node in the tree are returned.
}
\details{
Returns a list of lists: 'clade.children' contains all the children of each clade,
 'clade.node' contains the clade number of each clade in clade.children.
}
\examples{
# example.var <- exampleFun (test.data)
}
\references{
No references yet!
\url{https://github.com/DomBennett/MoreTreeTools}.
}
\keyword{ecology}
\keyword{evolution}
\keyword{manipulation}
\keyword{palaeontology}
\keyword{phylogeny}

