% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get-methods.R
\name{getEdgeStats}
\alias{getEdgeStats}
\title{Get edge statistics}
\usage{
getEdgeStats(tree, edges = "all", ignore.tips = NULL)
}
\arguments{
\item{tree}{phylogenetic tree (ape class)}

\item{edges}{numeric vector of edges, default 'all'}

\item{ignore.tips}{tips to ignore when counting, default NULL}
}
\description{
Return dataframe of first node of edge, second node of edge,
their ages and the number of children descending from the edge.
}
\details{
Assumes tree is rooted, bifurcating and time callibrated.

Use ignore.tips to specify tips of tree to ignore when counting
children.
}
\examples{
tree <- rtree (100)
edge.stats <- getEdgeStats (tree)
}
\references{
No references yet!
\url{https://github.com/DomBennett/MoreTreeTools}.
}
\keyword{ecology}
\keyword{evolution}
\keyword{manipulation}
\keyword{palaeontology}
\keyword{phylogeny}

