% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-methods.R
\name{getExtant}
\alias{getExtant}
\title{Return extant tips}
\usage{
getExtant(tree, tol = 1e-08)
}
\arguments{
\item{tree}{phylogenetic tree (ape class)}

\item{tol}{tolerance, below this is considered 0}
}
\description{
Return all tip labels whose tip age is 0.
}
\details{
No details
}
\examples{
# example.var <- exampleFun (test.data)
}
\references{
No references yet!
\url{https://github.com/DomBennett/MoreTreeTools}.
}
\keyword{ecology}
\keyword{evolution}
\keyword{manipulation}
\keyword{palaeontology}
\keyword{phylogeny}

