% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-methods.R
\name{getNodes}
\alias{getNodes}
\title{Get parent nodes}
\usage{
getNodes(tree, node)
}
\arguments{
\item{tree}{phylogenetic tree (ape class)}

\item{node}{number indicating node}
}
\description{
Return all nodes that connect the specified node to the root
}
\details{
No details
}
\examples{
# example.var <- exampleFun (test.data)
}
\references{
No references yet!
\url{https://github.com/DomBennett/MoreTreeTools}.
}
\keyword{ecology}
\keyword{evolution}
\keyword{manipulation}
\keyword{palaeontology}
\keyword{phylogeny}

