% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get-methods.R
\name{getNodeStats}
\alias{getNodeStats}
\title{Get node statistics}
\usage{
getNodeStats(tree, nodes = "all", ignore.tips = NULL)
}
\arguments{
\item{tree}{phylogenetic tree (ape class)}

\item{nodes}{numeric vector of nodes, default 'all'}

\item{ignore.tips}{tips to ignore when counting, default NULL}
}
\description{
Return dataframe of number of children, age and
previous connecting node of every node in tree.
}
\details{
Assumes tree is rooted, bifurcating and time callibrated.

Use ignore.tips to specify tips of tree to ignore when counting
children.
}
\examples{
tree <- rtree (100)
node.stats <- getNodeStats (tree)
}
\references{
No references yet!
\url{https://github.com/DomBennett/MoreTreeTools}.
}
\keyword{ecology}
\keyword{evolution}
\keyword{manipulation}
\keyword{palaeontology}
\keyword{phylogeny}

