% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-methods.R
\name{getOutgroup}
\alias{getOutgroup}
\title{Get outgroup in a tree from tips}
\usage{
getOutgroup(tree, tips)
}
\arguments{
\item{tree}{phylogenetic tree (ape class)}

\item{tips}{vector of tip labels}
}
\description{
Return the outgroup tip of a tree from a vector of tips given.
}
\details{
Note, if polytomies occur in the tree, outgroup maybe a vector.
}
\examples{
# example.var <- exampleFun (test.data)
}
\references{
No references yet!
\url{https://github.com/DomBennett/MoreTreeTools}.
}
\keyword{ecology}
\keyword{evolution}
\keyword{manipulation}
\keyword{palaeontology}
\keyword{phylogeny}

