% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-methods.R
\name{getParent}
\alias{getParent}
\title{Get parent node}
\usage{
getParent(tree, node = NULL, tips = NULL, edges = NULL)
}
\arguments{
\item{tree}{phylogenetic tree (ape class)}

\item{node}{vector of number indicating node}

\item{tips}{vector of tip labels or numbers}

\item{edges}{vector of edge labels or edge numbers}
}
\description{
Return parent node of nodes, edges or a vector of tips
}
\references{
No references yet!
\url{https://github.com/DomBennett/MoreTreeTools}.
}
\keyword{ecology}
\keyword{evolution}
\keyword{manipulation}
\keyword{palaeontology}
\keyword{phylogeny}

