\name{getSister}
\alias{getSister}
\title{Return sister node}
\usage{
getSister(tree, node)
}
\arguments{
  \item{tree}{phylogenetic tree (ape class)}

  \item{node}{number indicating node}
}
\description{
Return sister node of node given
}
\details{
No details
}
\examples{
# example.var <- exampleFun (test.data)
}
\references{
No references yet!
\url{https://github.com/DomBennett/MoreTreeTools}.
}
\keyword{ecology}
\keyword{evolution}
\keyword{manipulation}
\keyword{palaeontology}
\keyword{phylogeny}

