% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get-methods.R
\name{getSize}
\alias{getSize}
\title{Get the size of a phylogenetic tree}
\usage{
getSize(tree, type = c("ntips", "pd", "rtt"))
}
\arguments{
\item{tree}{phylogenetic tree (ape class)}

\item{type}{type of size to be calcualted: \code{ntips} (default), total branch length
\code{pd} or root to tip distance \code{rtt}}
}
\description{
Return either the number of tips, the total branch length
or the root to tip distance of a phylogenetic tree
}
\details{
I often need to quickly work out the size of a tree. This often means
writing \code{length (tree$tip.label)} over and over... but with this function
you can simply write \code{getSize(tree)}, that's a saving of 10 characters
(no sniffing matter for the hard core programmer) plus it looks neater. Not
to mention saving your brain time too.
}
\references{
No references yet!
\url{https://github.com/DomBennett/MoreTreeTools}.
}
\keyword{ecology}
\keyword{evolution}
\keyword{manipulation}
\keyword{palaeontology}
\keyword{phylogeny}

