% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get-methods.R
\name{getTreeStats}
\alias{getTreeStats}
\title{Return large list of trees stats by node}
\usage{
getTreeStats(tree)
}
\description{
Returns a list object of nodes with information
on each node -- see details.
}
\details{
Many functions require the navigation of a tree e.g.
finding all descending nodes, counting number of children by node
etc. \code{MoreTreeTools} has many functions for performing these
tasks, but in cases where these function need to be run multiple
times it can be more efficient to generate this information beforehand.

A list object is returned, with stats for each node. Stats
generated are:
\itemize{
  \item All nodes from root to node, \code{ascend.nodes}
  \item All edges from root to edge, \code{ascend.edges}
  \item Direct previous node connecting to node, \code{prev.nodes}
  \item Direct previous edges connecting to node, \code{prev.edges}
  \item Tips descending from node, \code{children}
  \item Number of tips descending from node, \code{n.children}
  \item Age of node, \code{age}
  \item Phylogenetic diversity of node, \code{pd}
}

Note, this function assumes trees are rooted. If no branch lengths are
given, the function sets all branch lengths to 1. The function is
recursive, for very large trees (100,000 + tips) a stack overflow warning
may appear. Update \code{options(expressions=)} to change the stack limit.
}
\examples{
# get hominoids tree
data ('hominoids')
# plot with labels
plot (hominoids);nodelabels();edgelabels();axisPhylo();
# generate tree.stats
tree.stats <- getTreeStats (hominoids)
# corroborate with plot
tree.stats[[19]][['prev.edges']]  # the previous edge for node 19 is 2
tree.stats[[20]][['children']]  # all the 'great' apes descend from node 20
}

