% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/calc-methods.R
\name{mapNames}
\alias{mapNames}
\title{Map names to a tree}
\usage{
mapNames(tree, names, fuzzy = TRUE, stats = FALSE, datasource = 4)
}
\arguments{
\item{tree}{phylogenetic tree (ape class)}

\item{names}{vector of names of tips to be extracted}

\item{fuzzy}{boolean, if true will search Global Names Resolver online}

\item{stats}{boolean, if true returns list object of extracted tree and extraction statistics}

\item{datasource}{GNR datasource ID, default 4 for NCBI}
}
\description{
Map names to a tree with the option of searching online
to perform fuzzy-name matching.
}
\details{
This function firsts matches names to tip labels in tree, if not all names
are present in the tree and 'fuzzy' is true, it will then search the online taxonomic
names resolution service Global Names Resolver. Taxonomic IDs are then matched
between tip labels and names, if names are still not mapped to tree, names are mapped
to tips using random mapping. For example, if an unmapped name is a member of the same
genus as several species in the tree, the name will be mapped to the tree with any one
of these at random. In cases where an unmapped name can only be added to the tree by adding
an extra tip, then a new tip is added at random to the pendant edge of a suitable tip at
a random time point (it assumes the tree is time callibrated) -- see examples.

In cases where a large proportion of the tips are mapped using random placement, it may
be best to generate a distribution of trees that represent the range of possible phylogenetic
relationships for the names and tree given.

If stats is true, the function will return a list of the extracted tree and a data frame
giving the proportions of names mapped to the tree, names mapped by exact string matching,
fuzzy matching and random taxonomic group placement. Of those names that are mapped using
random taxonomic placement, the mean rank number is returned. For the NCBI taxonomy, these
rank numbers are roughly equivalent to:

    Rank         Number

    superkingdom    2

    kingdom         4

    phylum          9

    subphylum       10

    superclass      12

    class           18

    superorder      21

    order           22

    suborder        23

    infraorder      24

    parvorder       25

    superfamily     26

    family          27

    subfamily       28

    genus           29

    species         30

These numbers depend on datasource and taxonomic group; some groups have more ranks than others.

If the mean rank number is very high, it is likely that the tree on which the names have been
mapped does not represent them very well.
}
\examples{
# bring in the catarrhines data
data ('catarrhines')
# we want to map these names to the catarrhines tree
names <- c ('Homo sapiens', 'Pongo pygmaeus', 'Gorilla gorila', 'Pan troglodytes',
            'Homo erectus', 'Homo neanderthalensis', 'Hylobates')
# 4 of the names are already in the tree (one has a spelling mistake) but the other 3
# do not exist and will be mapped randomly based on resolved taxonomic lineages of the
# names already in the tree and the names to be added.
hominid.map <- mapNames (tree=catarrhines, names=names, stats=TRUE, fuzzy=TRUE)
plot (hominid.map$tree)
print (hominid.map$stats)  # all names were mapped, 4 fuzzily on average at the genus level
}
\references{
No references yet!
\url{https://github.com/DomBennett/MoreTreeTools}.
}
\keyword{ecology}
\keyword{evolution}
\keyword{manipulation}
\keyword{palaeontology}
\keyword{phylogeny}

