% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map-methods.R
\name{mapNamesPreDownload}
\alias{mapNamesPreDownload}
\title{Download names for mapNames()}
\usage{
mapNamesPreDownload(names, datasource = 4)
}
\arguments{
\item{names}{vector of names to be resolved and all names in tree to be mapped}

\item{datasource}{GNR datasource ID, default 4 for NCBI}

\item{tree}{phylogenetic tree (ape class)}
}
\description{
Run this function before running mapNames() when
mapping multiple sets of names.
}
\details{
mapNames() will try to minimise the number of searches it
needs to do to map names given to a phylogenetic tree. If you're running
mapNames() multiple times for different vectors of names, it may be faster
to resolve names via this function and then pass these to mapNames()
}
\examples{
# bring in the catarrhines data
data ('catarrhines')
names1 <- c ('Homo sapiens', 'Pongo pygmaeus', 'Gorilla gorila', 'Pan troglodytes',
             'Homo erectus', 'Homo neanderthalensis', 'Hylobates')
names2 <- c ('Hylobates', 'Homo', 'Macca')
resolve.list <- mapNamesPreDownload (names=c (names1, names2, catarrhines$tip.label),
                                     datasource=4)
map1 <- mapNames (tree=catarrhines, names=names1, resolve.list=resolve.list)
map2 <- mapNames (tree=catarrhines, names=names2, resolve.list=resolve.list)
plot (map1)
plot (map2)
}
\references{
No references yet!
\url{https://github.com/DomBennett/MoreTreeTools}.
}
\keyword{ecology}
\keyword{evolution}
\keyword{manipulation}
\keyword{palaeontology}
\keyword{phylogeny}

