\name{nodeDescendants}
\alias{nodeDescendants}
\title{Return descendant species from a node}
\usage{
nodeDescendants(phylo, node, display = FALSE)
}
\arguments{
  \item{phylo}{phylogenetic tree (ape class)}

  \item{node}{number indicating node}

  \item{display}{if true, a tree will be plotted with
  specifed node highlighted}
}
\description{
Return all tip labels that descend from a specifed node.
}
\details{
No details
}
\examples{
# example.var <- exampleFun (test.data)
}
\references{
Me, I (2014) My wonderful paper, All about me, 1:1:1
\url{test.url}.
}
\keyword{manip}

