% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map-methods.R
\name{pinNames}
\alias{pinNames}
\title{Pin new tips to time-callibrated tree}
\usage{
pinNames(tree, names, lineages, min.ages, max.ages, fuzzy = TRUE,
  datasource = 4, iterations = 1, resolve.list = NULL)
}
\arguments{
\item{tree}{phylogenetic tree (ape class)}
}
\description{
Take a list of names with taxonomic lineages and age ranges, and pin them on
to existing time-callibrated tree.
}
\details{
This function ...
}
\examples{
# bring in the catarrhines data
}
\references{
No references yet!
\url{https://github.com/DomBennett/MoreTreeTools}.
}
\keyword{ecology}
\keyword{evolution}
\keyword{manipulation}
\keyword{palaeontology}
\keyword{phylogeny}

