% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/clade-methods.R
\name{plotClades}
\alias{plotClades}
\title{Plot clade success through time}
\usage{
plotClades(clades, k = 3, clade.stats = NULL, i = NULL)
}
\arguments{
\item{clades}{}

\item{k}{number of clades to plot}

\item{clade.stats}{stats calculated using calcCladeStats (optional)}

\item{i}{indexes of user selected of clades to plot (optional)}

\item{tree}{phylogenetic tree (ape class)}
}
\description{
Return a geom_object of clade success through time.
}
\details{
This function will produce a geom_object() of the top k
clades by size through time. It takes a dataframe of clade successes
through time produced with getCladeSuccess().

By default the function will select only the biggest k clades by total
size. Alternatively, the user can select the clades of interest by
specifying their column indexes in the clades dataframe.

The user can calculate clade stats inside or outside the function.
}
\examples{
# example.var <- exampleFun (test.data)
}
\references{
No references yet!
\url{https://github.com/DomBennett/MoreTreeTools}.
}
\keyword{ecology}
\keyword{evolution}
\keyword{manipulation}
\keyword{palaeontology}
\keyword{phylogeny}

