% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manip-methods.R
\name{reBalance}
\alias{reBalance}
\title{Change balance of tree}
\usage{
reBalance(tree, steps)
}
\arguments{
\item{tree}{phylogenetic tree (ape class)}

\item{steps}{number of tip switches to perform}
}
\description{
Return a tree with balance changed.
}
\details{
The functions switches tips in tree to make the tree
more or less balanced. If steps is positive, the tree will be
made more balanced else more unbalanced. Steps determines the
number of tip switches in a balanced or unbalanced direction.
}
\examples{
# Make an unbalanced tree balanced
n <- 16
tree <- compute.brlen (stree (n, 'left'))
plot (tree, show.tip.label=FALSE, edge.width=2, main='Unbalanced')
# +11 steps to make tree balanced
tree <- reBalance (tree, 11)
plot (tree, show.tip.label=FALSE, edge.width=2, main='Rebalanced')
}
\references{
No references yet!
\url{https://github.com/DomBennett/MoreTreeTools}.
}
\keyword{ecology}
\keyword{evolution}
\keyword{manipulation}
\keyword{palaeontology}
\keyword{phylogeny}

