% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/manip-methods.R
\name{reGravitise}
\alias{reGravitise}
\title{Change gravity of tree}
\usage{
reGravitise(tree, factor)
}
\arguments{
\item{tree}{phylogenetic tree (ape class)}

\item{factor}{proportion (-1 to +1) by which gravity is changed,
positive values increase gravity, negative values decrease it.}
}
\description{
Return a tree with gravity changed by factor.
}
\details{
The functions works by adding and removing branch length
between parent and child branches. The amount by which the branch
lengths are changed is determined by factor.
}
\examples{
# Create balanced tree, plot positively and negatively reGravitised
n <- 16
par (mfrow = c (3, 1), mar = c (1, 1, 1, 1))
tree <- compute.brlen (stree (n, 'balanced'))
tree$root.edge <- 0.1  # add a root edge
plot (tree, show.tip.label=FALSE, edge.width=2, root.edge=TRUE,
      main='Original')
plot (reGravitise (tree, factor = -0.9), show.tip.label=FALSE,
      edge.width=2, root.edge=TRUE, main='Reduced gravity')
plot (reGravitise (tree, factor = 0.9), show.tip.label=FALSE,
      edge.width=2, root.edge=TRUE, main='Increased gravity')
}
\references{
No references yet!
\url{https://github.com/DomBennett/MoreTreeTools}.
}
\keyword{ecology}
\keyword{evolution}
\keyword{manipulation}
\keyword{palaeontology}
\keyword{phylogeny}

