% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/sim-methods.R
\name{runEDBMM}
\alias{runEDBMM}
\title{Evolutionary Distinctness Biased Markov Model (EDBMM)}
\usage{
runEDBMM(birth, death, stop.at, stop.by = c("n", "t"), seed.tree = NULL,
  bias = c("FP", "ES", "PE"), sig = -1, eps = -1, record = FALSE,
  sample.at = stop.at * 0.1, fossils = FALSE, max.iteration = 10000,
  progress.bar = "none")
}
\arguments{
\item{birth}{how many births per unit of branch length}

\item{death}{how many deaths per unit of branch length}

\item{stop.at}{what value to stop at}

\item{stop.by}{what to stop at time or taxa}

\item{seed.tree}{tree to start model with, else a random tree of 2 tips}

\item{bias}{what type of ED bias to use FP or PE}

\item{sig}{power modifier of ED speciation bias}

\item{eps}{power modifier of ED extinction bias}

\item{record}{return list of trees at specified sample points}

\item{fossils}{keep fossils in tree (T/F)}

\item{progress.bar}{plyr progress bar, record only}

\item{tree}{phylogenetic tree (ape class)}

\item{max.iterations}{maximum number of iterations to prevent inf looping}
}
\description{
Run simulation of tree growth using the Evolutionary
Distinctness Biased Markov Model (EDBMM).
}
\details{
None yet!
}
\examples{
# example.var <- exampleFun (test.data)
}
\references{
No references yet!
\url{https://github.com/DomBennett/MoreTreeTools}.
}
\keyword{ecology}
\keyword{evolution}
\keyword{manipulation}
\keyword{palaeontology}
\keyword{phylogeny}

