% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/server-methods.R
\name{taxaResolve}
\alias{taxaResolve}
\title{Resolve taxonmic names online}
\usage{
taxaResolve(names, batch = 100, datasource = 4, genus = TRUE)
}
\arguments{
\item{names}{vector of names}

\item{batch}{size of the batches to be queried}

\item{datasource}{id number of the}

\item{genus}{boolean, if true will search against GNR with just the genus
name for names that failed to resovle using the full species name}
}
\description{
Resolve taxonomic names via the Global Names Resolver.
}
\details{
Returns dataframe containing GNR metadata for each name wames
that cannot be resolved are returned as NA. Various datasources are
available, see http://resolver.globalnames.biodinfo.org/data_sources for a
list and IDs. Default is 4 for NCBI.
}

