\name{testclades}
\alias{testclades}
\docType{data}
\title{Uniform and arching clade richnesses through time}
\description{
  Load clade successes as generated by \code{getCladeSuccess()} for simulated
  data. \code{unif.clades} contains 100 clades where each success is uniform
  through time. \code{arch.clades} contains 100 clades where each experiences
  a rise and fall through time. The simulated data is used to test\code{plotClades()}.
}
\usage{
   data (testclades)
}
\format{
  \code{testclades} contains two \code{data.frame}s:
  \code{unif.clades} and \code{arch.clades}
}
\examples{
data (testclades) # load objects
# plot
plotClades (clades=unif.clades, cids=1:ncol (unif.clades), merge=TRUE)
plotClades (clades=arch.clades, cids=1:ncol (arch.clades), merge=TRUE)
}
\keyword{datasets}
