% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/experimental-methods.R
\name{treeplot}
\alias{treeplot}
\title{Plot tree through time}
\usage{
treeplot(tree, edge.cols = NULL, edge.sizes = NULL, legend.title = "")
}
\arguments{
\item{tree}{}
}
\description{
Return a geom_object() of a phylogenetic tree through time.
Requires tree given to be rooted, time callibrated and bifurcating.
}
\details{
No details
}
\examples{
tree <- rtree (50)  # random tree with fossils
plot (tree)  # ape plot
treeplot (tree)  # treeplot
}

