% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/clade-methods.R
\name{calcCladeStats}
\alias{calcCladeStats}
\title{Calculate clade stats}
\usage{
calcCladeStats(clades)
}
\arguments{
\item{clades}{dataframe where columns are clades and rows are numbers
of descendents by time step as returned by getCladeSuccess()}
}
\description{
Return a dataframe of stats for dataframe produced from getCladeSuccess().
}
\details{
The function calculates the following statistics for each clade:

tot.size: the total number of tips in the clade through time

max.size: the maximum number of tips in the clade in one time step

time.span: the number of time steps the clade existed

start: the starting time step

end: the ending time step

cm: the centre of mass of the clade

cg: the centre of gyration of the clade
}
\examples{
library (MoreTreeTools)
# simulate an EDBMM tree (this will take time!)
trees <- runEDBMM (birth=2, death=1, stop.at=200, record=TRUE,
                   progress.bar='time')
# get raw number of descendents by node for trees
clades <- getCladeSuccess (trees)
# get independent number of descendents by node
clades.ind <- getCladeSuccess (trees, ind=TRUE)
# calculate stats by clade
clade.stats <- calcCladeStats (clades)
clade.ind.stats <- calcCladeStats (clades.ind)
# plot the trajectories of the 5 biggest clades
plotClades (clades=clades, clade.stats=clade.stats, k=5)
# plot clades through time for all clades that went extinct
# by the end of the simulation (there are 11 time steps in the simulation)
selection <- which (clade.stats$end < 11 &
                      clade.stats$max.size > 2)
plotClades (clades=clades, i=selection)
plotClades (clades=clades, i=selection, merge=TRUE)
# compare with ind clades
selection <- which (clade.ind.stats$end < 11 &
                      clade.ind.stats$max.size > 2)
plotClades (clades=clades.ind, i=selection)
plotClades (clades=clades.ind, i=selection, merge=TRUE)
# it is difficult to detect a difference in clade shape
# with so few time steps. Instead let's load some generated
# test clades
load ('testclades')
# this loads 'unif.clades' and 'arch.clades'
plotClades (clades=unif.clades, i=1:ncol (unif.clades), merge=TRUE)
plotClades (clades=arch.clades, i=1:ncol (arch.clades), merge=TRUE)
# now we can see a difference
}
\references{
Gould, S. J., Gilinsky, N. L., & German, R. Z. (1987). Asymmetry of
lineages and the direction of evolutionary time. Science (New York, N.Y.),
236(4807), 1437–1441.
}

