% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get-methods.R
\name{getAge}
\alias{getAge}
\title{Return age of node}
\usage{
getAge(tree, node = "all")
}
\arguments{
\item{tree}{phylogenetic tree (ape class)}

\item{node}{number indicating node (default 'all')}
}
\description{
Return age of tip or internal node
 (so long as the tree provided is time calibrated with branch lengths).
}
\details{
First calculates the root to tip distance, then calculates
node age by subtracting this distance from the root to node distance.
If \code{node} equals 'all', will return a dataframe for all nodes.
}
\examples{
# example.var <- exampleFun (test.data)
}
\references{
No references yet!
\url{https://github.com/DomBennett/MoreTreeTools}.
}
\keyword{ecology}
\keyword{evolution}
\keyword{manipulation}
\keyword{palaeontology}
\keyword{phylogeny}

