% Generated by roxygen2 (4.0.1): do not edit by hand
\name{getChildren}
\alias{getChildren}
\title{Return descendant species from a node}
\usage{
getChildren(tree, node, display = FALSE)
}
\arguments{
\item{tree}{phylogenetic tree (ape class)}

\item{node}{number indicating node}

\item{display}{if true, a tree will be plotted with specifed node highlighted}
}
\description{
Return all tip labels that descend from a specifed node.
}
\details{
No details
}
\examples{
# example.var <- exampleFun (test.data)
}
\references{
No references yet!
\url{https://github.com/DomBennett/MoreTreeTools}.
}
\keyword{ecology}
\keyword{evolution}
\keyword{manipulation}
\keyword{palaeontology}
\keyword{phylogeny}

