% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/clade-methods.R
\name{getCladeSuccess}
\alias{getCladeSuccess}
\title{Clade success through time}
\usage{
getCladeSuccess(trees, time.intervals = 1:length(trees))
}
\arguments{
\item{trees}{}

\item{time.intervals}{vector of time represented by each tree in trees}

\item{tree}{phylogenetic tree (ape class)}
}
\description{
Return a dataframe of clade tip number from a list
of trees.
}
\details{
The function takes a list of trees and counts the success
of each clade, returning a dataframe of number of tips for each
tree for each named clade. It requires nodes to be labeled in
the 'node.label' slot of the Phylo class.

Takes list of trees generated from runEDBMM(record=TRUE).

Note, a clade in this function is defined as any node with two
descending branches each with extant tips.
}
\examples{
# example.var <- exampleFun (test.data)
}
\references{
No references yet!
\url{https://github.com/DomBennett/MoreTreeTools}.
}
\keyword{ecology}
\keyword{evolution}
\keyword{manipulation}
\keyword{palaeontology}
\keyword{phylogeny}

