% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/clade-methods.R
\name{getCladeSuccess}
\alias{getCladeSuccess}
\title{Clade success through time}
\usage{
getCladeSuccess(trees, ind = FALSE, count.extincts = FALSE,
  time.intervals = 1:length(trees))
}
\arguments{
\item{trees}{}

\item{ind}{boolean, use independent counting? Deafult FALSE}

\item{count.extincts}{boolean, count extinct and extant tips? Default FALSE}

\item{time.intervals}{vector of time represented by each tree in trees}

\item{tree}{phylogenetic tree (ape class)}
}
\description{
Return a dataframe of clade tip number from a list
of trees.
}
\details{
The function takes a list of trees and counts the success
of each clade, returning a dataframe of number of tips for each
tree for each named clade. It requires nodes to be labeled in
the 'node.label' slot of the Phylo class.

Takes list of trees generated from runEDBMM(record=TRUE).

Setting 'ind' (or independent) as TRUE will implement a counting
method that attempts to count descendents unique to the clade by
preventing double counting in decsendent clades. Independent
counting works by limiting the maximum clade count to the minimum
of the descendant counts of the two daughter clades. If one of the
daughter clades continues to speciate while the other goes extinct
then the parent clade of the two will be limited to 0 rather than
the size of a single daughter clade.

'time.intervals' is a vector of time steps represented by each tree.
Default is 1 to length of number of trees.

Setting 'count.extincts' to TRUE will count both extinct and extant
tips through time.
}
\examples{
library (MoreTreeTools)
# simulate an EDBMM tree (this will take time!)
trees <- runEDBMM (birth=2, death=1, stop.at=200, record=TRUE,
                   progress.bar='time')
# get raw number of descendents by node for trees
clades <- getCladeSuccess (trees)
# get independent number of descendents by node
clades.ind <- getCladeSuccess (trees, ind=TRUE)
# calculate stats by clade
clade.stats <- calcCladeStats (clades)
clade.ind.stats <- calcCladeStats (clades.ind)
# plot the trajectories of the 5 biggest clades
plotClades (clades=clades, clade.stats=clade.stats, k=5)
# plot clades through time for all clades that went extinct
# by the end of the simulation (there are 11 time steps in the simulation)
selection <- which (clade.stats$end < 11 &
                      clade.stats$max.size > 2)
plotClades (clades=clades, i=selection)
plotClades (clades=clades, i=selection, merge=TRUE)
# compare with ind clades
selection <- which (clade.ind.stats$end < 11 &
                      clade.ind.stats$max.size > 2)
plotClades (clades=clades.ind, i=selection)
plotClades (clades=clades.ind, i=selection, merge=TRUE)
# it is difficult to detect a difference in clade shape
# with so few time steps. Instead let's load some generated
# test clades
data ('testclades')
# this loads 'unif.clades' and 'arch.clades'
plotClades (clades=unif.clades, i=1:ncol (unif.clades), merge=TRUE)
plotClades (clades=arch.clades, i=1:ncol (arch.clades), merge=TRUE)
# now we can see a difference
}
\references{
No references yet!
\url{https://github.com/DomBennett/MoreTreeTools}.
}
\keyword{ecology}
\keyword{evolution}
\keyword{manipulation}
\keyword{palaeontology}
\keyword{phylogeny}

