% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/get-methods.R
\name{getParent}
\alias{getParent}
\title{Get parent node}
\usage{
getParent(tree, node = NULL, tips = NULL)
}
\arguments{
\item{tree}{phylogenetic tree (ape class)}

\item{node}{number indicating node}

\item{tips}{tip labels or numbers}
}
\description{
Return parent node of a node or a vector of tips
}
\references{
No references yet!
\url{https://github.com/DomBennett/MoreTreeTools}.
}
\keyword{ecology}
\keyword{evolution}
\keyword{manipulation}
\keyword{palaeontology}
\keyword{phylogeny}

