% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/map-methods.R
\name{mapNames}
\alias{mapNames}
\title{Map names to a tree}
\usage{
mapNames(tree, names, fuzzy = TRUE, datasource = 4, iterations = 1)
}
\arguments{
\item{tree}{phylogenetic tree (ape class)}

\item{names}{vector of names of tips to be extracted}

\item{fuzzy}{boolean, if true will search Global Names Resolver online}

\item{datasource}{GNR datasource ID, default 4 for NCBI}

\item{iterations}{how many times to repeat?}
}
\description{
Map names to a tree with the option of searching online
to perform fuzzy-name matching.
}
\details{
This function firsts matches names to tip labels in tree, if not all names
are present in the tree and 'fuzzy' is true, it will then search the online taxonomic
names resolution service Global Names Resolver. Taxonomic IDs are then matched
between tip labels and names, if names are still not mapped to tree, names are mapped
to tips using random mapping. For example, if an unmapped name is a member of the same
genus as several species in the tree, the name will be mapped to the tree with any one
of these at random. In cases where an unmapped name can only be added to the tree by adding
an extra tip, then a new tip is added at random to the pendant edge of a suitable tip at
a random time point (it assumes the tree is time callibrated) -- see examples.

In cases where a large proportion of the tips are mapped using random placement, it may
be best to generate a distribution of trees that represent the range of possible phylogenetic
relationships for the names and tree given. If iterations is greater than 1, instead of a
single tree, a multiPhylo object of length iterations is returned representing this
range of possible mappings.

For non-time-callibrated or trees without branch lengths, the function using taxonomic
distance for random placement and returns a tree without branch lengths.
}
\examples{
# bring in the catarrhines data
data ('catarrhines')
# we want to map these names to the catarrhines tree
names <- c ('Homo sapiens', 'Pongo pygmaeus', 'Gorilla gorila', 'Pan troglodytes',
            'Homo erectus', 'Homo neanderthalensis', 'Hylobates')
# 4 of the names are already in the tree (one has a spelling mistake) but the other 3
# do not exist and will be mapped randomly based on resolved taxonomic lineages of the
# names already in the tree and the names to be added.
hominid.map <- mapNames (tree=catarrhines, names=names, fuzzy=TRUE)
plot (hominid.map)
}
\references{
No references yet!
\url{https://github.com/DomBennett/MoreTreeTools}.
}
\keyword{ecology}
\keyword{evolution}
\keyword{manipulation}
\keyword{palaeontology}
\keyword{phylogeny}

