% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/clade-methods.R
\name{plotClades}
\alias{plotClades}
\title{Plot clade success through time}
\usage{
plotClades(clades, k = 3, clade.stats = NULL, i = NULL, legend = FALSE,
  merge = FALSE)
}
\arguments{
\item{clades}{}

\item{k}{top number of clades to plot}

\item{clade.stats}{stats calculated using calcCladeStats (optional)}

\item{i}{indexes of user selected of clades to plot (optional)}

\item{legend}{boolean, display legend? Default FALSE}

\item{merge}{boolean, plot normalised clade trajectory? Default FALSE}
}
\description{
Return a geom_object of clade success through time.
}
\details{
This function will return a geom_object for plotting
the trajectories of clade success through time from a clades
dataframe generated by getCladeSuccess().

By default, the largest three by maximum size clades will be plotted.
This number can be changed with the k parameter. If using top k, for
dataframes of large sizes, it is computationally more efficient to
generate the clade.stats dataframe before calling this function with
calcCladeStats().

The user can also choose to plot the trajectories of specific clades
in the clades dataframe using 'i' e.g. to show only clades that
have gone extinct or only clades of a certain size.

The user can also merge the clade trajectories by setting 'merge' as
TRUE. This will take the trajectory of each clade and normalise their
time and number of descendents by putting them onto the same scale.
Normalised time is calculated as the time when the clade appeared (0)
to when it went it extinct (1). Normalise number of descendents is
rescaled by setting the smallest size of the clade to 0 and the
largest size to 1.
}
\examples{
library (MoreTreeTools)
# simulate an EDBMM tree (this will take time!)
trees <- runEDBMM (birth=2, death=1, stop.at=200, record=TRUE,
                   progress.bar='time')
# get raw number of descendents by node for trees
clades <- getCladeSuccess (trees)
# get independent number of descendents by node
clades.ind <- getCladeSuccess (trees, ind=TRUE)
# calculate stats by clade
clade.stats <- calcCladeStats (clades)
clade.ind.stats <- calcCladeStats (clades.ind)
# plot the trajectories of the 5 biggest clades
plotClades (clades=clades, clade.stats=clade.stats, k=5)
# plot clades through time for all clades that went extinct
# by the end of the simulation (there are 11 time steps in the simulation)
selection <- which (clade.stats$end < 11 &
                      clade.stats$max.size > 2)
plotClades (clades=clades, i=selection)
plotClades (clades=clades, i=selection, merge=TRUE)
# compare with ind clades
selection <- which (clade.ind.stats$end < 11 &
                      clade.ind.stats$max.size > 2)
plotClades (clades=clades.ind, i=selection)
plotClades (clades=clades.ind, i=selection, merge=TRUE)
# it is difficult to detect a difference in clade shape
# with so few time steps. Instead let's load some generated
# test clades
load ('testclades')
# this loads 'unif.clades' and 'arch.clades'
plotClades (clades=unif.clades, i=1:ncol (unif.clades), merge=TRUE)
plotClades (clades=arch.clades, i=1:ncol (arch.clades), merge=TRUE)
# now we can see a difference
}

