% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/clade-methods.R
\name{plotClades}
\alias{plotClades}
\title{Plot clade success through time}
\usage{
plotClades(clades, k = 3, clade.stats = NULL, i = NULL, legend = FALSE,
  merge = FALSE)
}
\arguments{
\item{clades}{}

\item{k}{number of clades to plot}

\item{clade.stats}{stats calculated using calcCladeStats (optional)}

\item{i}{indexes of user selected of clades to plot (optional)}

\item{tree}{phylogenetic tree (ape class)}
}
\description{
Return a geom_object of clade success through time.
}
\details{
This function will produce a geom_object() of the top k
clades by size through time. It takes a dataframe of clade successes
through time produced with getCladeSuccess().

By default the function will select only the biggest k clades by total
size. Alternatively, the user can select the clades of interest by
specifying their column indexes in the clades dataframe.

The user can calculate clade stats inside or outside the function.
}
\examples{
library (MoreTreeTools)
# simulate an EDBMM tree (this will take time!)
trees <- runEDBMM (birth=2, death=1, stop.at=200, record=TRUE,
                   progress.bar='time')
# get raw number of descendents by node for trees
clades <- getCladeSuccess (trees)
# get independent number of descendents by node
clades.ind <- getCladeSuccess (trees, ind=TRUE)
# calculate stats by clade
clade.stats <- calcCladeStats (clades)
clade.ind.stats <- calcCladeStats (clades.ind)
# plot the trajectories of the 5 biggest clades
plotClades (clades=clades, clade.stats=clade.stats, k=5)
# plot clades through time for all clades that went extinct
# by the end of the simulation (there are 11 time steps in the simulation)
selection <- which (clade.stats$end < 11 &
                      clade.stats$max.size > 2)
plotClades (clades=clades, i=selection)
plotClades (clades=clades, i=selection, merge=TRUE)
# compare with ind clades
selection <- which (clade.ind.stats$end < 11 &
                      clade.ind.stats$max.size > 2)
plotClades (clades=clades.ind, i=selection)
plotClades (clades=clades.ind, i=selection, merge=TRUE)
# it is difficult to detect a difference in clade shape
# with so few time steps. Instead let's load some generated
# test clades
load ('testclades')
# this loads 'unif.clades' and 'arch.clades'
plotClades (clades=unif.clades, i=1:ncol (unif.clades), merge=TRUE)
plotClades (clades=arch.clades, i=1:ncol (arch.clades), merge=TRUE)
# now we can see a difference
}
\references{
No references yet!
\url{https://github.com/DomBennett/MoreTreeTools}.
}
\keyword{ecology}
\keyword{evolution}
\keyword{manipulation}
\keyword{palaeontology}
\keyword{phylogeny}

