% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clade-methods.R
\name{plotClades}
\alias{plotClades}
\title{Plot clade success through time}
\usage{
plotClades(clades, N = 3, clade.stats = NULL, cids = NULL,
  legend = FALSE, merge = FALSE)
}
\arguments{
\item{N}{top number of clades to plot}

\item{clade.stats}{stats calculated using calcCladeStats (optional)}

\item{cids}{clade ids of user selected clades to plot (optional)}

\item{legend}{boolean, display legend? Default FALSE}

\item{merge}{boolean, plot normalised clade trajectory? Default FALSE}
}
\description{
Return a geom_object of clade success through time.
}
\details{
This function will return a geom_object for plotting
the trajectories of clade success through time from a clades
dataframe generated by getCladeSuccess().

By default, the largest three by maximum size clades will be plotted.
This number can be changed with the N parameter. If using top N, for
dataframes of large sizes, it is computationally more efficient to
generate the clade.stats dataframe before calling this function with
calcCladeStats().

The user can also choose to plot the trajectories of specific clades
in the clades dataframe using 'cids' e.g. to show only clades that
have gone extinct or only clades of a certain size.

The user can also merge the clade trajectories by setting 'merge' as
TRUE. This will take the trajectory of each clade and normalise their
time and number of descendents by putting them onto the same scale.
Normalised time is calculated as the time when the clade appeared (0)
to when it went it extinct (1). Normalise number of descendents is
rescaled by setting the smallest size of the clade to 0 and the 
largest size to 1. Black line indicates the mean while red dashed
indicate the SD.
}

