% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/manip-methods.R
\name{reLoad}
\alias{reLoad}
\title{Change loading of tree}
\usage{
reLoad(tree, factor)
}
\arguments{
\item{tree}{phylogenetic tree (ape class)}

\item{factor}{proportion (-1 to +1) by which loading is changed,
positive values increase loading, negative values decrease it.}
}
\description{
Return a tree with loading changed by factor.
}
\details{
The functions works by adding and removing branch length
between parent and children branches. The amount by which the branch
lengths are changed is determiend by factor.
}
\examples{
# Create balanced tree, plot positively and negatively reloaded
n <- 16
par (mfrow = c (2, 2), mar = c (1, 1, 1, 1))
tree <- compute.brlen (stree (n, 'balanced'))
tree$root.edge <- 0.1  # add a root edge
plot (reLoad (tree, factor = -0.9), show.tip.label=FALSE,
      edge.width=2, root.edge=TRUE)
plot (reLoad (tree, factor = 0.9), show.tip.label=FALSE,
      edge.width=2, root.edge=TRUE)
}
\references{
No references yet!
\url{https://github.com/DomBennett/MoreTreeTools}.
}
\keyword{ecology}
\keyword{evolution}
\keyword{manipulation}
\keyword{palaeontology}
\keyword{phylogeny}

