% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build-methods.R
\name{removeTip}
\alias{removeTip}
\title{Remove a tip from a phylogenetic tree}
\usage{
removeTip(tree, tips, preserve.age)
}
\arguments{
\item{tree}{phylogenetic tree (ape class)}

\item{tips}{vector of tip names}

\item{preserve.age}{maintain the root to tip distance of the tree}
}
\description{
Remove tips from a phylogenetic tree
}
\details{
Based on \code{ape}'s \code{drop.tip} but with
\code{preserve.age}. This will ensure
the root to tip distance of the tree is maintained by adding length
to next immediate internal edge whenever the removal of a tip
leads to a reduction in a tree's age. If root node is lost, then
lost edge length is added to \code{root.edge} slot.
}
\examples{
tree <- rtree (10)
par (mfrow=c (1,3))
plot (tree)
plot (removeTip (tree, tips=c ('t1', 't2', 't3'), preserve.age=FALSE))
plot (removeTip (tree, tips=c ('t1', 't2', 't3'), preserve.age=TRUE), root.edge=TRUE)
}
\references{
No references yet!
\url{https://github.com/DomBennett/MoreTreeTools}.
}
\keyword{ecology}
\keyword{evolution}
\keyword{manipulation}
\keyword{palaeontology}
\keyword{phylogeny}

